python-bot/pyproject.toml
[tool.poetry]
name = "python-bot"
version = "0.1.0"
description = "Product Visor Python bot"
authors = ["ru.dmitartur <bot@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
requests = "^2.31.0"
pyyaml = "^6.0.1"
aiokafka = "^0.10.0"
pydantic = "^2.8.2"
pydantic-settings = "^2.4.0"
prometheus-client = "^0.19.0"
aiogram = "^3.4.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


